Moved firebase api to frontend/.env
