+++ frontend/firebase.js
@@ -1,11 +1,6 @@
-// Import the functions you need from the SDKs you need
 import { initializeApp } from "firebase/app";
 import { getAuth } from "firebase/auth";
 
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
 const firebaseConfig = {
   apiKey: "AIzaSyDPkSTmR28jzMny0fqDKpRADi0znQu_MbI",
   authDomain: "skincare-recommender-3a679.firebaseapp.com",
@@ -15,9 +10,7 @@ const firebaseConfig = {
   appId: "1:1039328606924:web:e0d24ddbc9457567fde069"
 };
 
-// Initialize Firebase
 const app = initializeApp(firebaseConfig);
 
-// Initialize Firebase Authentication and get a reference to the service
 export const auth = getAuth(app);
 export default app;
