+++ frontend/src/pages/login.jsx
@@ -1,7 +1,7 @@
 {/* This is the login. */}
 import React, { useState } from "react";
 import { signInWithEmailAndPassword } from "firebase/auth";
-import { auth } from "../firebase";
+import { auth } from "../../firebase";
 import { NavLink, useNavigate } from "react-router-dom";
 
 const Login = () => {
