+++ frontend/src/pages/signup.jsx
@@ -2,7 +2,7 @@
 import React, {useState} from 'react';
 import { NavLink, useNavigate } from 'react-router-dom';
 import {  createUserWithEmailAndPassword  } from 'firebase/auth';
-import { auth } from '../firebase';
+import { auth } from '../../firebase';
 
 const Signup = () => {
     const navigate = useNavigate();
